refactor: break up `models/group/group.go` further
